requirements.txt
opencv-python
Pillow
numpy
pandas
